requirements.txt
bandit>=1.7.4
black>=23.1.0
dj-rest-auth==2.2.7
# django
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
# docs
drf-yasg==1.21.4
# tools
flake8>=6.0.0
# wsgi
gunicorn==20.1.0
isort>=5.12.0
Markdown==3.4.1
# testing
model_bakery=>1.10.1
psycopg2-binary==2.9.5
# redis
redis==4.5.0
# ros
roslibpy==1.4.1
